requirements.txt
aiohttp>=3.9,<4
botbuilder-core>=4.14,<5
botbuilder-integration-aiohttp>=4.14,<5
gunicorn>=21,<23
azure-identity>=1.15,<2
azure-storage-blob>=12.19,<13
openai>=1.30,<2
msal>=1.28,<2
apscheduler>=3.10,<4
